# Remove line packing from tiles, slightly improve tile culling