--- a/sparse_strips/vello_cpu/snapshots/eo_filling_missing_anti_aliasing.png
+++ b/sparse_strips/vello_cpu/snapshots/eo_filling_missing_anti_aliasing.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b79e38722c2ad12febaecdf8d4dcb57421ce01a132ce8bbd1dfebb1a79afd91e
-size 219
+oid sha256:18fa5af2e5279deb475be420e2772cd9a82e77dc025c108bd61f8bc569938501
+size 213