--- a/sparse_strips/vello_cpu/snapshots/filling_evenodd_rule.png
+++ b/sparse_strips/vello_cpu/snapshots/filling_evenodd_rule.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b6c3921c0acad8dca99d955aeb0564422df1c3d18bd366d556f9d55951e27b23
-size 832
+oid sha256:c30cefd5e417136aa65bfdc90cc841514e2bafce765914991c3f5d4cac2c0439
+size 840