--- a/sparse_strips/vello_cpu/snapshots/filling_nonzero_rule.png
+++ b/sparse_strips/vello_cpu/snapshots/filling_nonzero_rule.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f1440086e64002849d92e51292e22a6b78c91b970082cec7a7784c9aafec1386
-size 717
+oid sha256:33e8a40b36dc176dac9a7aa6b7ed86c2cf4e72f15223ea56e6bc8021d23551ec
+size 723