… sparse_strips/vello_cpu/snapshots/filling_unclosed_path_2.png
… sparse_strips/vello_cpu/snapshots/incorrect_filling_1.png
… sparse_strips/vello_cpu/snapshots/incorrect_filling_3.png
… sparse_strips/vello_cpu/snapshots/rectangle_above_viewport.png
… sparse_strips/vello_cpu/snapshots/rectangle_left_of_viewport.png
… sparse_strips/vello_cpu/snapshots/stroked_triangle.png
… sparse_strips/vello_cpu/snapshots/triangle_above_and_wider_than_viewport.png
… sparse_strips/vello_cpu/src/render.rs
… sparse_strips/vello_cpu/tests/basic.rs
… sparse_strips/vello_toy/src/debug.rs