This reduces the tile memory footprint and drops the requirement for tile x-coordinates to be signed. We may be able to drop the `TileIndex` indirection soon. 

Coarse winding calculation is performed in tile generation, like it more-or-less was before. In principle it requires just a single bit in the tile packing. Calculating the coarse winding could be performed in the strip generation as well, but I suspect it to be slightly more performant this way. I haven't accurately measured that.

This is one step towards what [Raph mentioned](https://xi.zulipchat.com/#narrow/channel/197075-gpu/topic/CPU.20sparse.20strip.20rendering.20to.20pixels/near/500019583). Quoting:

> pack x, y, line id, and winding number delta into a single u32 and sort that. This is a very tight packing with u32 and may overflow in some cases (very large number of lines per path), but I think should be very efficient. Then the determination of p0 and p1 within the tile happens after sorting. A major motivation for this change is that it can be done on GPU in hybrid modes, but the approach to parallelism also works with SIMD, I think.

Tiles above, to the right and below the viewport are now culled. Tiles to the left of the viewport are clamped but not culled yet. Currently strip generation has some code special-casing those tiles.

On my machine, paris-30k measures with #835 as the "before" and after the changes here as the following. (With 20 iterations, appears to be accurate to about ~0.2ms, but we should think about better benchmarking.)

```
Tile generation
    old: 24.394854ms
    new: 17.620834ms
Tile sorting
    old: 8.922622ms
    new: 8.146455ms
Strip generation
    old: 33.187645ms
    new: 31.825052ms
```