@tomcur 2025-03-10
Spot-checking some pixels, the changed tests all appear to just be the rounding of pixel coverage falling differently. I haven't checked too closely, but it's possible the area coverage calculation are by coincidence slightly more symmetrical now, check e.g. the left and right edges of `eo_filling_missing_anti_aliasing.png` before and after these changes.)