@tomcur 2025-03-11
To reduce some effort, tile x-coord data type clean up and packing the tile more tightly are best done together, I think.